requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
anyio==4.9.0
asgiref==3.9.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
boto3==1.39.3
botocore==1.39.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
dataclasses-json==0.6.7
distlib==0.3.9
Django==5.2.4
django-environ==0.12.0
django-extensions==4.1
django-microsoft-auth==3.0.1
dnspython==2.7.0
filelock==3.18.0
frozenlist==1.7.0
gql==3.5.3
graphql-core==3.2.6
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
marshmallow==3.26.1
multidict==6.6.3
mypy_extensions==1.1.0
numpy==2.3.1
oauthlib==3.3.1
packaging==25.0
pandas==2.3.0
platformdirs==4.3.8
propcache==0.3.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pymongo==4.13.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
s3transfer==0.13.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
taegis-sdk-python==1.4.8
typing-inspect==0.9.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
xlsxwriter==3.2.5
yarl==1.20.1

# Testing dependencies
pytest==8.0.0
pytest-django==4.8.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
freezegun==1.4.0
responses==0.25.0
openpyxl==3.1.2